requirements.txt
gradio==3.1.5          # Library for creating the interactive frontend
matplotlib==3.9.2      # For generating attention map visualizations
numpy==2.1.3           # Required for numerical operations
Pillow==9.5.0          # For image processing and manipulation
torch==2.5.1           # PyTorch for model training and inference
torchvision==0.20.1    # Provides CIFAR-10 dataset and data transformations
transformers==4.46.2   # Hugging Face library for pre-trained ViT models